requirements.txt
asgiref==3.2.7
Django==3.0.7
django-crispy-forms==1.9.1
gunicorn==20.0.4
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
